Use double for quantity instead of int

+++ Libraries/SPTarkov.Server.Core/Helpers/Dialogue/Commando/SptCommands/TraderCommand/TraderSptCommand.cs
@@ -50,7 +50,7 @@ public class TraderSptCommand(
 
         var trader = result.Groups["trader"].Captures.Count > 0 ? result.Groups["trader"].Captures[0].Value : null;
         var command = result.Groups["command"].Captures.Count > 0 ? result.Groups["command"].Captures[0].Value : null;
-        var quantity = int.Parse(result.Groups["command"].Captures.Count > 0 ? result.Groups["quantity"].Captures[0].Value : "0");
+        var quantity = double.Parse(result.Groups["command"].Captures.Count > 0 ? result.Groups["quantity"].Captures[0].Value : "0");
 
         var dbTrader = _traderHelper.GetTraderByNickName(trader);
         if (dbTrader == null)
@@ -109,7 +109,7 @@ public class TraderSptCommand(
         return request.DialogId;
     }
 
-    protected ProfileChangeEvent CreateProfileChangeEvent(ProfileChangeEventType profileChangeEventType, int quantity, string dbTraderId)
+    protected ProfileChangeEvent CreateProfileChangeEvent(ProfileChangeEventType profileChangeEventType, double quantity, string dbTraderId)
     {
         return new ProfileChangeEvent
         {
